+++ src/Client/Auth/AuthMethod.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace Addwiki\Mediawiki\Api\Client\Auth;
+
+interface AuthMethod {
+}