+++ RELEASENOTES.md
@@ -6,6 +6,9 @@
 - Typing added throughout
 - PSR4 namespacing. Now in `Addwiki\Mediawiki\Api`
 - Client classes, such as `MediaWikiApi`, moved to `Addwiki\Mediawiki\Api`
+- `ApiUser` deprecated in favour of `UserAndPassword` and `UserAndPasswordWithDomain`
+- `MediawikiApiInterface` `isLoggedin` renamed to `isLoggedIn`
+- `MediawikiApiInterface` and implementations will now only return a boolean from `isLoggedIn`
 
 ## Version 2.8 (16 February 2021)